device: remove default-unmanaged
Get rid of NM_UNMANAGED_DEFAULT and replace it for most parts
of NM_UNMANAGED_USER.

The user-flag can be modified via D-Bus API. Also, when the user
activates a connection, it automatically clears NM_UNMANAGED_USER.
This should for the most part give a similar behavior to the previous
NM_UNMANAGED_DEFAULT behavior.

However, when deactivating a device that was previously
NM_UNMANAGED_DEFAULT, we would no longer transition the device to
unmanaged-state.

https://bugzilla.gnome.org/show_bug.cgi?id=746566

Based-on-patch-by: Lubomir Rintel <lkundrak@v3.sk>