fixup! firewall: refactor callback handling in NMFirewallManager
Always take a reference to the manager. The manager should stay alive
as long as there are pending operations.

On shutdown, NMDevices are not destroyed. Thus, if we destroy the
firewall manager with pending operations, their callback will execute
again on the device, which might be unexpected.
With this, we easily leak NMFirewallManager on shutdown.