device: fix checking for vlan parent in component_added()
Commit cd3df12c8f8ed6c868c12bc4e7fe6ba162dafc5b reused the
virtual function component_added() to notify the vlan device
about a possibly new parent.

This reuse of the virtual function for another purpose is confusing.
Clean that up by splitting the implementation and add a new
virtual function nm_device_notify_new_device_added() which gets
(only implemented by NMDeviceVlan).

But more important, when calling nm_device_notify_new_device_added()
self might not realized itself. Thus, we need to gracefully return
in that case. I still think the handling of NMDeviceVlan in this
situation is broken. It should more closely monitor for the master
device. Still, for now, this fixes the issue.